Move cells and columns to TokenLoansContent
