--- src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/LoansTokenActiveTable.tsx
@@ -1,5 +1,0 @@
-const LoansTokenActiveTable = () => {
-  return <></>
-}
-
-export default LoansTokenActiveTable
