--- src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/index.ts
@@ -1,1 +1,0 @@
-export * from './cells'
