--- src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LoansTokenActiveTable'
